+++ config-example.json
@@ -498,6 +498,10 @@
             "track5": {
                 "file": "track5.mp3",
                 "volume": 0.3
+            },
+            "track6": {
+                "file": "track6.mp3",
+                "volume": 0.5
             }
         }
     },
@@ -563,8 +567,8 @@
             "size2": "map_size2.json",
             "size3": "map_size3.json",
             "size4": "map_size4.json",
-            "test": "map_test.json",
-            "test2": "map_test2.json"
+            "tutorial 1": "map_tutorial_1.json",
+            "tutorial 2": "map_tutorial_2.json"
         }
     },
     "regions": {
