+++ maps/map_tutorial_2.json
@@ -30,6 +30,12 @@
             "person_1": null
         },
         {
+            "person_2": 65,
+            "transport": true,
+            "subType": "metro",
+            "at": 65
+        },
+        {
             "person_2": null
         }
     ],
@@ -44,11 +50,11 @@
     "difficulty": 1,
     "max": 1,
     "total": 3,
-    "score": 0,
+    "score": 1,
     "completion": {
-        "completed": false
+        "completed": true
     },
-    "audio": "track1",
+    "audio": "track6",
     "backgrounds": {
         "layer 1": [
             246,
