+++ sprites/entity.py
@@ -1,8 +1,10 @@
 import pygame
 import math
 import abc
-from config import YELLOW, WHITE, BLACK
-from utils import overrides, vec, getScale
+import person as PERSON
+from pygame.locals import BLEND_MIN
+from config import YELLOW, WHITE, BLACK, HOVERGREY
+from utils import overrides, vec, getScale, getMousePos
 from sprite import Sprite
 
 
@@ -99,7 +101,7 @@ class MouseClick(Sprite):
             self.spriteRenderer.belowEntities), self.target, infinite=True)
 
         # We want to clear the target to allow for double clicks (on taxi)
-        if self.direction == "left":
+        if self.direction == "left" and isinstance(self.target, PERSON.Person):
             self.clickManager.setTarget(None)
 
     def setNextClick(self):
@@ -113,6 +115,26 @@ class MouseClick(Sprite):
             return
 
         direction = "right" if self.direction == "left" else "left"
+
+        # If the next (node) is on a layer below another node, 
+        # we want to show clicking on the indicator
+        if len(self.next.getBelow()) > 0:
+            for indicator in self.next.getBelow():
+                # We want to find the indicator for the current target.
+                if indicator.getTarget() != self.next:
+                    continue
+
+                # No need to click on indicator if we're already
+                # on the correct layer.
+                if (int(indicator.getTarget().getConnectionType()[-1])
+                        == self.spriteRenderer.getCurrentLayer()):
+                    continue
+
+                MouseClick(
+                    self.groups, indicator, self.tutorialManager,
+                    [self.clickManager], next=self.next)
+                return
+
         MouseClick(
             self.groups, self.next, self.tutorialManager,
             [self.clickManager], direction)
@@ -134,8 +156,12 @@ class MouseClick(Sprite):
 
     @overrides(Sprite)
     def events(self):
-        if (self.direction == "left"
-                and self.clickManager.getTarget() == self.target):
+        if (self.direction == "left" and (
+                self.clickManager.getTarget() == self.target
+                or (
+                    isinstance(self.target, BelowIndicator)
+                    and int(self.target.getTarget().getConnectionType()[-1])
+                    == self.spriteRenderer.getCurrentLayer()))):
             self.setNextClick()
 
         # Since you can only right click on a node,
@@ -446,3 +472,68 @@ class StatusIndicator(Sprite):
         if (self.target.getStatusValue() - 1) != self.currentState:
             self.dirty = True
             self.currentState = self.target.getStatusValue() - 1
+
+
+class BelowIndicator(Sprite):
+    def __init__(self, groups, currentNode, target):
+        super().__init__(currentNode.spriteRenderer, groups, [])
+        self.currentNode = currentNode
+        self.target = target
+
+        self.width, self.height = 11.5, 11.5
+        self.offset = vec(21, 19 + (
+            len(self.currentNode.getAbove()) * self.width))
+        self.pos = self.currentNode.pos + self.offset
+
+    def getTarget(self):
+        return self.target
+
+    def __render(self):
+        self.dirty = False
+
+        self.image = self.game.imageLoader.getImage(
+            self.target.getImages()[0], (
+                self.width * self.spriteRenderer.getFixedScale(),
+                self.height * self.spriteRenderer.getFixedScale()))
+        self.rect = self.image.get_rect()
+
+        self.rect.topleft = self.getTopLeft(self)
+
+    @overrides(Sprite)
+    def makeSurface(self):
+        if self.dirty or self.image is None:
+            self.__render()
+
+    @overrides(Sprite)
+    def events(self):
+        if self.game.mainMenu.getOpen():
+            return
+        mx, my = getMousePos(self.game)
+
+        # Click event; take the player to the layer below that is
+        # being indicated.
+        if (self.rect.collidepoint((mx, my))
+                and self.game.clickManager.getClicked()
+                and self.game.clickManager.getMouseOver() == self):
+            self.game.clickManager.setClicked(False)
+            self.spriteRenderer.showLayer(
+                int(self.target.getConnectionType()[-1]))
+
+        # Hover over event.
+        elif (self.rect.collidepoint((mx, my)) and not self.mouseOver
+                and self.game.clickManager.isTop(self)):
+            self.mouseOver = True
+            self.image.fill(HOVERGREY, special_flags=BLEND_MIN)
+            self.game.clickManager.setMouseOver(self)
+            self.dirty = False
+
+        # Hover out event.
+        elif not self.rect.collidepoint((mx, my)) and self.mouseOver:
+            self.mouseOver = False
+            self.game.clickManager.setMouseOver(None)
+            self.dirty = True
+
+    @overrides(Sprite)
+    def update(self):
+        if not self.dirty:
+            self.events()
